Implement stub version
It'd be too dangerous to test moving money between pots with an actual
account, so extracted an interface from the monzo api so that we can
have a stub implementation that we can use for testing purposes.

There's no guarantee the stub behaves like the real thing, but we can't
really run the same tests against a real account to make sure is the
case. However, the idea is to use the api in very simple way, so this
stub should be enough.