Command to distribute funds into pots
Using the bank abstraction, this command takes funds from a source pot
and distribute them into the deposit pots.

As the source pot could have a different amount every month, define a
remainder pot that takes the funds that were not distributed, so that
the main account is unaffected - all the funds that get withdrawn from
the source pot get distributed.

The remainder can be used for savings, so defined a minimum amount that
should make it into the pod. This way, you could ensure you're saving
enough, or the command fails. In the case your income is greater than
usual, then you just end up with more funds in the pot, which should
actually be a good thing, if the pot was about savings.