Validate pots exist before attempting to distribute funds
Running the command against my actual account failed because some pots
had spaces at the end of their name, so they couldn't be found.

Therefore, we need to validate they exist before starting to make
movements,

However, some movements were made before the pot with the name issue was
found, so part of the distribution happened, and the command couldn't be
re-run without manually moving the funds back to the source.

Validating that all the pots included in the manifest exist will prevent
this issue from happening.