Trim pot names when parsing monzo response
Running the distribute funds command against my actual account failed
because some pots had spaces at the end of their name, so they couldn't
be found.

This issue was considered a bug on monzo's part, as having leading or
trailing spaces shouldn't be allowed. For that reason, the change is
made in the monzo implementation rather than in the command. A
consequence of that is that we can't really test it, as we can't run
tests against an actual monzo account. This was considered acceptable
given the simplicity of the problem.