Implement the funds distributor http application
Use the funds distributor contract to drive the implementation.
Unfortunately, it can't use a real version of the bank abstraction, as
that would move real money on every run.

Implementing the http adapter highlighted a weakness in the design, as
the errors are signaled using a generic exception, which forces the
adapter to catch it and expect that it represents a bad request.

We've stuck to the kotlinx serialisation library, which meant we
couldn't re-use the manifest type, unless we wanted to pollute the model
with serialisation annotations. This might be a good thing in the end,
as you're forced to have a http specific type rather than poke with the
serialiser. The only issue is naming, as you naturally tend to want to
use the same name for the http types.

In the main, we now expect an env variable for the monzo api access
token. The oauth flow is to be implemented in a future iteration.